Fix type bugs
